Zoom should be a double

+++ src/AzureMapsControl.Components/Animations/Options/MoveAlongPathAnimationOptions.cs
@@ -68,6 +68,6 @@
         /// <summary>
         /// A fixed zoom level to snap the map to on each animation frame.
         /// </summary>
-        public int? Zoom { get; set; }
+        public double? Zoom { get; set; }
     }
 }
